push printf erevi final

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: aadyan <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/01 17:51:54 by aadyan            #+#    #+#             */
-/*   Updated: 2024/11/04 14:01:53 by aadyan           ###   ########.fr       */
+/*   Updated: 2024/11/04 16:53:59 by aadyan           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,11 +15,12 @@
 
 # include <stdarg.h>
 # include <unistd.h>
+# include <stdio.h>
 
 int		ft_putnbr(int nb);
 int		ft_putchar(char c);
 int		check_print(char **str, va_list arg);
-int		dec_to_hex(unsigned int nb, int lower_case);
+int		dec_to_hex(unsigned long long nb, int lower_case);
 int		print_pointer(void *ptr);
 int		ft_putstr(char *str);
 int		ft_put_unsigned_nbr(unsigned int nb);
